Overhaul of tests for numeric types: * Significantly improved tests for numeric types using boundary value analysis such that we cover all supported number types inside and beyond their bounds * Improved tests for enum types * Fixed a small bug that didn't allow for mapping primitive Java doubles using boxed Java Doubles and was assuming they were floats * Removed code from unimplemented feature that was meant to allow for enums to be mapped using longs in addition to ints, some code was existing but never did anything as a dedicated EnumLongConverter was required, this functionality can be added later if needed

+++ src/main/java/jnr/ffi/provider/jffi/ToNativeOp.java
@@ -58,8 +58,8 @@ abstract class ToNativeOp {
         }
         m.put(float.class, new Float32(float.class));
         m.put(Float.class, new Float32(Float.class));
-        m.put(double.class, new Float64(float.class));
-        m.put(Double.class, new Float64(Float.class));
+        m.put(double.class, new Float64(double.class));
+        m.put(Double.class, new Float64(Double.class));
         m.put(Address.class, new AddressOp());
 
         operations = Collections.unmodifiableMap(m);
